+++ llvm/include/llvm/DebugInfo/CodeView/DebugCrossImpSubsection.h
@@ -36,7 +36,7 @@ template <> struct VarStreamArrayExtractor<codeview::CrossModuleImportItem> {
 public:
   using ContextType = void;
 
-  Error operator()(BinaryStreamRef Stream, uint32_t &Len,
+  Error operator()(BinaryStreamRef Stream, size_t &Len,
                    codeview::CrossModuleImportItem &Item);
 };
