+++ llvm/include/llvm/DebugInfo/CodeView/DebugSubsectionRecord.h
@@ -72,7 +72,7 @@ private:
 } // end namespace codeview
 
 template <> struct VarStreamArrayExtractor<codeview::DebugSubsectionRecord> {
-  Error operator()(BinaryStreamRef Stream, uint32_t &Length,
+  Error operator()(BinaryStreamRef Stream, size_t &Length,
                    codeview::DebugSubsectionRecord &Info) {
     // FIXME: We need to pass the container type through to this function.  In
     // practice this isn't super important since the subsection header describes
