+++ llvm/lib/DebugInfo/CodeView/DebugInlineeLinesSubsection.cpp
@@ -21,7 +21,7 @@ using namespace llvm;
 using namespace llvm::codeview;
 
 Error VarStreamArrayExtractor<InlineeSourceLine>::
-operator()(BinaryStreamRef Stream, uint32_t &Len, InlineeSourceLine &Item) {
+operator()(BinaryStreamRef Stream, size_t &Len, InlineeSourceLine &Item) {
   BinaryStreamReader Reader(Stream);
 
   if (auto EC = Reader.readObject(Item.Header))
